requirements.txt
numpy
ctranslate2==4.5.0
fugashi==1.3.3
jaconv==0.4.0
mss==10.0.0
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
paddleocr==2.9.1
Pillow==11.1.0
PySide6==6.8.1
PySide6_Addons==6.8.1
PySide6_Essentials==6.8.1
sentencepiece==0.2.0
torch-kmeans==0.2.0
transformers==4.46.2
unidic-lite==1.0.8
deepl==1.20.0
googletrans == 4.0.2